requirements.txt
crewai>=0.102.0,<1.0.0
python-dotenv>=1.0.1,<2.0.0
